refactor(IT-Cotato#301): TemporaryPost Appender, Deleter -> Manager로 통합
